Remove future date for event
Jekyll will not display posts (including posts with the "event"
category) if they are set for a future date.

When creating a post for an event which is occurring in the future, the
best practice is to simply not add a "Date" field in the frontmatter.
The filename should be based on the day the event is posted (so it shows
up in correct chronological order); Other info (like the title and body
text) can refer to the event's time and location.